Move SortOrder enum to grafana/data
