+++ public/app/core/utils/richHistory.test.ts
@@ -10,7 +10,7 @@ import {
   filterAndSortQueries,
 } from './richHistory';
 import store from 'app/core/store';
-import { SortOrder } from './explore';
+import { SortOrder } from './richHistory';
 import { dateTime, DataQuery } from '@grafana/data';
 
 const mock: any = {
