+++ public/app/features/explore/RichHistory/RichHistoryQueriesTab.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import { ExploreId } from '../../../types/explore';
-import { SortOrder } from 'app/core/utils/explore';
+import { SortOrder } from 'app/core/utils/richHistory';
 import { RichHistoryQueriesTab, Props } from './RichHistoryQueriesTab';
 import { Slider } from '@grafana/ui';
