+++ public/app/features/explore/RichHistory/RichHistoryStarredTab.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import { ExploreId } from '../../../types/explore';
-import { SortOrder } from 'app/core/utils/explore';
+import { SortOrder } from 'app/core/utils/richHistory';
 import { RichHistoryStarredTab, Props } from './RichHistoryStarredTab';
 
 jest.mock('../state/selectors', () => ({ getExploreDatasources: jest.fn() }));
