+++ public/app/plugins/panel/logs/types.ts
@@ -1,8 +1,8 @@
-import { SortOrder } from 'app/core/utils/explore';
+import { LogsSortOrder } from '@grafana/data';
 
 export interface Options {
   showLabels: boolean;
   showTime: boolean;
   wrapLogMessage: boolean;
-  sortOrder: SortOrder;
+  sortOrder: LogsSortOrder;
 }
